server-ng config
