+++ core/configs/src/server_config/server.rs
@@ -30,10 +30,7 @@ use err_trail::ErrContext;
 use figment::providers::{Format, Toml};
 use figment::value::Dict;
 use figment::{Metadata, Profile, Provider};
-use iggy_common::{
-    IggyByteSize, IggyDuration, MemoryPoolConfigOther, Validatable,
-    sharding::{MAX_PARTITIONS, MAX_STREAMS, MAX_TOPICS},
-};
+use iggy_common::{IggyByteSize, IggyDuration, MemoryPoolConfigOther, Validatable};
 use serde::{Deserialize, Serialize};
 use serde_with::DisplayFromStr;
 use serde_with::serde_as;
@@ -48,8 +45,6 @@ const DEFAULT_CONFIG_PATH: &str = "core/server/config.toml";
 pub struct ServerConfig {
     pub consumer_group: ConsumerGroupConfig,
     pub data_maintenance: DataMaintenanceConfig,
-    #[serde(default)]
-    pub extra: ExtraConfig,
     pub message_saver: MessageSaverConfig,
     pub personal_access_token: PersonalAccessTokenConfig,
     pub heartbeat: HeartbeatConfig,
@@ -60,29 +55,6 @@ pub struct ServerConfig {
     pub websocket: WebSocketConfig,
     pub telemetry: TelemetryConfig,
     pub cluster: ClusterConfig,
-}
-
-// TODO: Rename this to something more sensible, once we figure out all of the extra configuration we need.
-#[derive(Debug, Default, Deserialize, Serialize, Clone, ConfigEnv)]
-pub struct ExtraConfig {
-    pub namespace: NamespaceConfig,
-}
-
-#[derive(Debug, Deserialize, Serialize, Clone, ConfigEnv)]
-pub struct NamespaceConfig {
-    pub max_streams: usize,
-    pub max_topics: usize,
-    pub max_partitions: usize,
-}
-
-impl Default for NamespaceConfig {
-    fn default() -> Self {
-        Self {
-            max_streams: MAX_STREAMS,
-            max_topics: MAX_TOPICS,
-            max_partitions: MAX_PARTITIONS,
-        }
-    }
 }
 
 /// Configuration for the memory pool.
